+++ PlayAsOne.py
@@ -5,7 +5,10 @@ import Tkinter as tk
 import threading
 import time
 import ttk
-
+from flask import Flask, request, flash, url_for, redirect, \
+    render_template, abort, send_from_directory, jsonify
+from flask.ext.socketio import SocketIO, emit
+from json import loads, dumps
 
 class PlayAsOne:
 
@@ -15,6 +18,9 @@ class PlayAsOne:
         self.running = False
         self.window_active = False
         self.window_region = None
+
+        threading.Thread(target=socketio.run, args=(app,)).start()
+        self.gui.mainloop()
 
     def start(self):
         self.running = True
@@ -191,3 +197,73 @@ class GUI(tk.Tk):
             self.screenshot_button.config(
                 state='normal', text='Retake Screenshot')
         threading.Thread(target=func).start()
+
+app = Flask(__name__)
+socketio = SocketIO(app, async_mode='eventlet')
+user_count = 0
+users = {}
+democracy = []
+
+
+@app.route("/")
+def hello():
+    print("User entered")
+    return render_template('index.html')
+
+
+@socketio.on('connect', namespace="/")
+def test_connect():
+    print('test ran')
+
+
+@socketio.on("add user", namespace="/")
+def handle_add_user(username):
+    global user_count
+    user_count += 1
+    user = username
+    print user
+    users[user] = {'input_count': 0}
+    emit("initialize", {
+         'input_type': gui_server.get_input_mode(), 'mode': gui_server.get_mode()})
+
+
+def handle_chaos(user_input):
+    gui_server.execute_input(user_input)
+
+
+def handle_democracy(user_input):
+    global democracy
+    for demo_input in democracy:
+        if demo_input[0] == user_input['input']:
+            demo_input[1] += 1
+            break
+    democracy.append((user_input['input'], 1))
+
+
+def execute_democracy():
+    most_votes = ("", 0)
+    for user_input in democracy:
+        if user_input[1] > most_votes[1]:
+            most_votes = user_input
+    gui_server.execute_input(most_votes[0])
+
+
+@socketio.on('on disconnect', namespace="/")
+def handle_disconnect(json):
+    global user_count
+    user_count -= 1
+    user = loads(json)
+    users.pop(user['username'])
+
+
+@socketio.on("sendInput", namespace="/")
+def handle_input(json):
+    user_input = loads(json)
+    users[input['username']]['input_count'] += 1
+    if (gui_server.get_mode() == 'Chaos'):
+        handle_chaos(user_input)
+    elif (gui_server.get_mode() == 'Democracy'):
+        handle_democracy(user_input)
+
+if __name__ == '__main__':
+    gui_server = PlayAsOne()
